\name{distillateflow}
\alias{distillateflow}
\alias{distillate.flow}
\docType{data}
\title{The flow rate of distillate from the top of a distillation column.}
\description{
  These are actual data, taken 1 minute apart in time, of the flow rate leaving 
  the top of a continuous distillation column (data are from a 31 day period 
  in time).
}
\usage{data(distillateflow)}
\format{A data.frame containing 44,640 observations of 1 variable}
\source{Kevin Dunn, http://openmv.net}
\keyword{datasets}