\name{oildoe}
\alias{oilDOE}
\alias{oil.doe}
\docType{data}
\title{Industrial designed experiment to improve the volumetric heat capacity of a product.}
\description{
  Four materials: A, B, C and D are added in a blend to achieve a desired
  heat capacity, the response variable, y.
  
  The amounts were varied in a factorial manner for the 4 materials.
  
  The data are scaled and coded for confidentiality. All that may be disclosed is that
  variable C is either added ("Yes") or not added not added ("No").
}
\usage{data(pollutant)}
\format{A data.frame containing 19 observations of 5 variables (A, B, C, D, and the response, y)}
\source{Kevin Dunn, data from a confidential industrial source.}
\keyword{datasets}